Refactor app with love and care
Appointment information from the api was updated so it was referenced
by the slot. This commit updates the appointment to add a hash of
data to pass the user id and slot id on selection.
@jrhorn424 was able to take the slot id and user id and attach them
to the appointment creation properly. I really can't thank Jeff
enough for all the help that he's provided to me over the past couple of
days. This project is a huge deal for me and I was getting frustrated
with a number of ways that I had the project configured. With Jeff's
guidance and hard work we were able to change how this application works
and make what is the start of a beautiful project.